Added Homebrew / Linuxbrew support
A ruby script to install GTM via Homebrew or Linuxbrew has been added. It has been inspired by the ruby script for the original GTM.
Additionally, a README.md with information on how to download GTM via Homebrew or Linuxbrew and a link to the main repository has been provided.